AuthGuard implementation, covers AC defined in issue #6
