+++ src/app/modules/auth/auth.guard.ts
@@ -1,20 +1,33 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, CanActivateChild, CanLoad, Route, UrlSegment, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { CanActivate, CanActivateChild, CanLoad, Route, UrlSegment, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
 import { Observable } from 'rxjs';
+import { AuthService } from '../../services/auth.service';
+import { environment } from 'src/environments/environment.prod';
+import { ITokenUser } from './interfaces/itoken-user';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthGuard implements CanActivate, CanActivateChild, CanLoad {
+  constructor(private router: Router, private authService: AuthService) { }
+
   canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-    return true;
+      if (!this.authService.isLoggedIn()) {
+        this.router.navigate(['/login'], { queryParams: { returnUrl: state.url }});
+      }
+
+      return this.authService.isAdminRole();
   }
   canActivateChild(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-    return true;
+      if (!this.authService.isLoggedIn()) {
+        this.router.navigate(['/login'], { queryParams: { returnUrl: state.url }});
+      }
+
+      return this.authService.isAdminRole();
   }
   canLoad(
     route: Route,
