+++ src/app/modules/auth/interfaces/itoken-user.ts
@@ -1,0 +1,7 @@
+export interface ITokenUser {
+    firstName: string;
+    lastName: string;
+    roles: string[];
+    token: string;
+    userId: string;
+}
