+++ src/app/services/auth.service.ts
@@ -4,7 +4,8 @@ import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { environment } from 'src/environments/environment';
 import { Observable, throwError } from 'rxjs';
 import { catchError, tap, map } from 'rxjs/operators';
-import { IRegistrationFormData, ILoginFormData, ITokenUser } from '../interfaces/auth';
+import { IRegistrationFormData, ILoginFormData } from '../interfaces/auth';
+import { ITokenUser } from '../modules/auth/interfaces/itoken-user';
 import { LoggerService } from './logger.service';
 
 @Injectable({
@@ -15,6 +16,7 @@ export class AuthService {
   private jwtTokenName: string;
   private jwtToken: string;
   private localStorageUserKey: string;
+  private tokenUser: ITokenUser;
 
   constructor(private httpClient: HttpClient, 
     private jwtHelper: JwtHelperService, 
@@ -23,6 +25,7 @@ export class AuthService {
     this.jwtTokenName = environment.jwtTokenKey;
     this.localStorageUserKey = environment.localStorageUserKey;
     this.jwtToken = localStorage.getItem(this.jwtTokenName);
+    this.tokenUser = JSON.parse(localStorage.getItem(this.localStorageUserKey));
   }
 
   register(formData: IRegistrationFormData): Observable<any> {
@@ -42,16 +45,16 @@ export class AuthService {
       .pipe(
         tap(res => {
           if (res.token && res.success === true) {
-            let tokenUser:ITokenUser = { 
+            this.tokenUser = { 
               firstName: res.firstName, 
               lastName: res.lastName, 
               roles: res.roles, 
               token: res.token, 
               userId: res.userId };
 
-            this.jwtToken = tokenUser.token;
+            this.jwtToken = this.tokenUser.token;
             localStorage.setItem(this.jwtTokenName, this.jwtToken);
-            localStorage.setItem(this.localStorageUserKey, JSON.stringify(tokenUser));
+            localStorage.setItem(this.localStorageUserKey, JSON.stringify(this.tokenUser));
           } else {
             this.logger.log(res, 'AuthService.login [tap]');
           }
@@ -72,7 +75,20 @@ export class AuthService {
   }
 
   public isLoggedIn() {
-
     return this.jwtToken && !this.jwtHelper.isTokenExpired();
+  }
+
+  public isInRole(roleName: string): boolean {
+    if (!this.isLoggedIn()) return false;
+    
+    return this.tokenUser.roles.some((role) => role.toLowerCase() == roleName.toLowerCase());
+  }
+
+  public isAdminRole(): boolean {
+    return this.isInRole('admin');
+  }
+
+  public isUserRole(): boolean {
+    return this.isInRole('user');
   }
 }
